Hide Openrouter provider and add a README.  (#485)
## Summary
Hide Openrouter provider and add a README. Since the Openrouter provider
is not finalized yet, I'm putting it inside an internal directory so
people don't try to use it in a broken state. I'm also adding a README
for the project.

## How was it tested?
Ran tests.

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).