Upgrade the version of the OpenAI SDK we use to the latest stable v1.  (#487)
## Summary
Upgrade the version of the OpenAI SDK we use to the latest stable v1.
After this change, I'll update Axiom. But I need this updated AISDK in
order to be able to update Axiom. One of the downsides of working across
two repos.

## How was it tested?
I ran all the unit tests, and they passed. 

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).