[ai] Add StreamText with an example (#510)
## Summary
Adds StreamText and StreamTextStr to the top-level `ai` package. These
are the streaming counterparts to GenerateText and GenerateTextStr.

Adds an example showing the streaming case.

## How was it tested?
Ran the example successfully.

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).